Edge routing in dvi now works again :-) Fixed #1016
git-svn-id: svn://localhost:8080/caleydo_main/trunk@5696 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.view.dvi/src/org/caleydo/view/dvi/layout/edge/rendering/connectors/TopBundleConnector.java
@@ -1,19 +1,19 @@
 /*******************************************************************************
  * Caleydo - visualization for molecular biology - http://caleydo.org
- * 
+ *
  * Copyright(C) 2005, 2012 Graz University of Technology, Marc Streit, Alexander
  * Lex, Christian Partl, Johannes Kepler University Linz </p>
- * 
+ *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any later
  * version.
- * 
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
  * details.
- * 
+ *
  * You should have received a copy of the GNU General Public License along with
  * this program. If not, see <http://www.gnu.org/licenses/>
  *******************************************************************************/
@@ -86,11 +86,11 @@ public class TopBundleConnector extends ABundleConnector {
 		float bundlingPositionXOtherNode = calcXPositionOfBundlingPoint(otherNode,
 				commonTablePerspectives);
 
-		float deltaX = (float) (bundlingPositionXOtherNode - bundlingPositionX);
+		float deltaX = bundlingPositionXOtherNode - bundlingPositionX;
 		float ratioX = deltaX / viewFrustum.getWidth();
 
-		float edgeAnchorX = (float) bundlingPositionX + ratioX * node.getWidth() / 2.0f;
-		float edgeAnchorXOtherNode = (float) bundlingPositionXOtherNode - ratioX
+		float edgeAnchorX = bundlingPositionX + ratioX * node.getWidth() / 2.0f;
+		float edgeAnchorXOtherNode = bundlingPositionXOtherNode - ratioX
 				* otherNode.getWidth() / 2.0f;
 		float edgeAnchorY = (float) node.getBoundingBox().getMaxY()
 				+ pixelGLConverter
@@ -144,6 +144,7 @@ public class TopBundleConnector extends ABundleConnector {
 
 		// Point2D bandAnchorPoint1 = null;
 		// Point2D bandAnchorPoint2 = null;
+
 
 		calcBandAnchorPoints(isEnd1, bandPoints);
